+++ e2e/test/scenarios/dashboard/actions-on-dashboards.cy.spec.js
@@ -445,6 +445,117 @@ const MODEL_NAME = "Test Action Model";
             }
           });
         });
+
+        it("properly loads and updates date and time fields for implicit update actions", () => {
+          createModelFromTable(TEST_COLUMNS_TABLE);
+          cy.get("@modelId").then(id => {
+            createImplicitAction({
+              kind: "update",
+              model_id: id,
+            });
+          });
+
+          createDashboardWithActionButton({
+            actionName: "Update",
+            idFilter: true,
+          });
+
+          filterWidget().click();
+          addWidgetStringFilter("1");
+
+          clickHelper("Update");
+
+          cy.wait("@executePrefetch");
+
+          const oldRow = many_data_types_rows[0];
+          const newTime = "2020-01-10T01:35:55";
+
+          modal().within(() => {
+            changeValue({
+              fieldName: "Date",
+              fieldType: "date",
+              oldValue: oldRow.date,
+              newValue: newTime.slice(0, 10),
+            });
+
+            changeValue({
+              fieldName: "Datetime",
+              fieldType: "datetime-local",
+              oldValue: oldRow.datetime.replace(" ", "T"),
+              newValue: newTime,
+            });
+
+            changeValue({
+              fieldName: "Time",
+              fieldType: "time",
+              oldValue: oldRow.time,
+              newValue: newTime.slice(-8),
+            });
+
+            changeValue({
+              fieldName: "Timestamp",
+              fieldType: "datetime-local",
+              oldValue: oldRow.timestamp.replace(" ", "T"),
+              newValue: newTime,
+            });
+
+            // only postgres has timezone-aware columns
+            // the instance is in US/Pacific so it's -8 hours
+            if (dialect === "postgres") {
+              changeValue({
+                fieldName: "Datetimetz",
+                fieldType: "datetime-local",
+                oldValue: "2020-01-01T00:35:55",
+                newValue: newTime,
+              });
+
+              changeValue({
+                fieldName: "Timestamptz",
+                fieldType: "datetime-local",
+                oldValue: "2020-01-01T00:35:55",
+                newValue: newTime,
+              });
+            }
+
+            if (dialect === "mysql") {
+              changeValue({
+                fieldName: "Datetimetz",
+                fieldType: "datetime-local",
+                oldValue: oldRow.datetimeTZ.replace(" ", "T"),
+                newValue: newTime,
+              });
+
+              changeValue({
+                fieldName: "Timestamptz",
+                fieldType: "datetime-local",
+                oldValue: oldRow.timestampTZ.replace(" ", "T"),
+                newValue: newTime,
+              });
+            }
+            cy.button("Update").click();
+          });
+
+          cy.wait("@executeAPI");
+
+          queryWritableDB(
+            `SELECT * FROM ${TEST_COLUMNS_TABLE} WHERE id = 1`,
+            dialect,
+          ).then(result => {
+            const row = result.rows[0];
+
+            // the driver adds a time to this date so we have to use .include
+            expect(row.date).to.include(newTime.slice(0, 10));
+            expect(row.time).to.equal(newTime.slice(-8));
+
+            // metabase is smart and localizes these, so all of these are +8 hours
+            const newTimeAdjusted = newTime.replace("T01", "T09");
+            // we need to use .include because the driver adds milliseconds to the timestamp
+            expect(row.datetime).to.include(newTimeAdjusted);
+            expect(row.timestamp).to.include(newTimeAdjusted);
+            expect(row.datetimeTZ).to.include(newTimeAdjusted);
+            expect(row.timestampTZ).to.include(newTimeAdjusted);
+          });
+        });
       });
     },
   );
